chore: Move gem metadata logic into separate TU.
